Completely reworked queue and play function
