+++ music.js
@@ -1,35 +1,10 @@
 const tools = require('./tools.js'),
+    shared = require('./shared.js'),
     ytdl = require('ytdl-core'),
     streamOptions = { seek: 0, volume: 1 };
 
 let botVoiceChannel = null,
-    botConnection = null,
-    queue = [];
-
-let dispatcher;
-
-let sendMusicEmbed = (message, music, bot) => {
-    let time = {};
-    time = tools.getTimeFormat(music.duration);
-    message.channel.send('', {embed : {
-        color: 65399,
-        author: {
-            name: `BOT ${message.guild.name} : MODE DJ`,
-            icon_url: bot.user.avatarURL
-        },
-        title: 'Now playing :',
-        fields: [{
-            name: 'Titre : ',
-            value: `${music.title}`
-        }, {
-            name: 'Durée :',
-            value: `${time.hours} heures ${time.minutes} minutes et ${time.seconds} secondes`
-        }, {
-            name: 'Proposée par :',
-            value: `${music.author}`
-        }]
-    }});
-}
+    botConnection = null;
 
 module.exports = {
 /*
@@ -55,7 +30,7 @@ module.exports = {
  Command : !play [link]
 */
     handlePlay : (message, bot) => {
-        let tmp = queue[0] ? true : false;
+        let tmp = shared.musicQueues[message.guild.id].queue[0] ? true : false;
         let tab = message.content.split(' ');
         let music = {};
         if (!tab[1]) {
@@ -64,34 +39,21 @@ module.exports = {
         } else if (!botConnection) {
             message.channel.send('Il faut que je soit dans un channel vocal pour utilier cette commande');
             return ;
-        } else {
-            music.url = tab[1];
-            music.author = message.author.username;
-            ytdl.getInfo(tab[1])
-            .then(info => {
-                music.title = info.title;
-                music.duration = info.length_seconds;
-                if (!tmp)
-                    sendMusicEmbed(message, music, bot);
-                else
-                    message.channel.send(`\`${music.title}\` a été ajouté à la file par \`${music.author}\``)
-            }).catch(console.error());
-            if (!queue[0]) {
-                const stream = ytdl(music.url, {filter : 'audioonly'});
-                this.dispatcher = botConnection.playStream(stream, streamOptions);
-            }
-            this.dispatcher.on('end', (end) => {
-                console.log(end);
-                queue.shift();
-                if (queue[0]) {
-                    sendMusicEmbed(message, queue[0], bot);
-                    stream = ytdl(queue[0].url, {filter : 'audioonly'});
-                    this.dispatcher = botConnection.playStream(stream, streamOptions);
-                }
-            });
-            queue.push(music);
         }
-        message.delete();
+        music.url = tab[1];
+        music.author = message.author.username;
+        ytdl.getInfo(tab[1])
+        .then(info => {
+            music.title = info.title;
+            music.duration = info.length_seconds;
+            if (tmp)
+                message.channel.send(`\`${music.title}\` a été ajouté à la file par \`${music.author}\``)
+        }).catch(console.error());
+        if (!shared.musicQueues[message.guild.id])
+            musicQueues[message.guild.id] = { queue : [] };
+        shared.musicQueues[message.guild.id].queue.push(music);
+        tools.playMusic(botConnection, message);
+        //message.delete();
     },
 
 /*
@@ -122,8 +84,7 @@ module.exports = {
  Command : !skip
  */
     handleSkip : message => {
-        if (queue[0] && this.dispatcher) {
-            this.dispatcher.end();
-        }
+        if (shared.musicQueues[message.guild.id].dispatcher)
+            shared.musicQueues[message.guild.id].dispatcher.end();
     }
 }
