+++ shared.js
@@ -9,5 +9,6 @@ try {
 module.exports = {
     adminRoles : JSON.parse(adminlist_s),
     niggerTab : [],
-    killConfirm : false
+    killConfirm : false,
+    musicQueues : {}
 }
